pyproject.toml
[project]
name = "mkflashcards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "beautifulsoup4>=4.12.3",
    "google-generativeai>=0.8.2",
    "instructor>=1.5.0",
    "langchain-text-splitters>=0.3.0",
    "logfire[openai,requests,starlette]>=1.0.1",
    "markdown>=3.7",
    "openai>=1.51.0",
    "pydantic>=2.9.2",
    "python-fasthtml>=0.6.9",
    "requests>=2.32.3",
    "tiktoken>=0.8.0",
]
